Split DataManagerTest into DataManagerExceptionTest and DataManagerNoExceptionTest.
