+++ dataland-backend/src/test/kotlin/org/dataland/datalandbackend/utils/TestDataProvider.kt
@@ -4,10 +4,14 @@ import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.dataland.datalandbackend.entities.StoredCompanyEntity
 import org.dataland.datalandbackend.frameworks.lksg.model.LksgData
+import org.dataland.datalandbackend.model.DataType
+import org.dataland.datalandbackend.model.StorableDataset
 import org.dataland.datalandbackend.model.companies.CompanyInformation
 import org.dataland.datalandbackend.model.enums.company.IdentifierType
+import org.dataland.datalandbackend.services.CompanyAlterationManager
 import org.springframework.beans.factory.annotation.Autowired
 import java.io.File
+import java.time.Instant
 
 class TestDataProvider(
     @Autowired var objectMapper: ObjectMapper,
@@ -49,4 +53,20 @@ class TestDataProvider(
             false,
             null,
         )
+
+    fun addCompanyAndReturnStorableDatasetForIt(
+        companyAlterationManager: CompanyAlterationManager,
+        frameworkName: String,
+    ): StorableDataset {
+        val companyInformation = getCompanyInformation(1).first()
+        val companyId = companyAlterationManager.addCompany(companyInformation).companyId
+        return StorableDataset(
+            companyId,
+            DataType(frameworkName),
+            "USER_ID_OF_AN_UPLOADING_USER",
+            Instant.now().toEpochMilli(),
+            "",
+            "someData",
+        )
+    }
 }
